payments_flask/requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
razorpay==1.4.2
supabase==2.6.0
gunicorn==22.0.0
